api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
psycopg2-binary>=2.9.0
